requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Data manipulation and analysis
pandas>=1.3.0

# Machine learning utilities
scikit-learn>=1.0.0

# Graph processing
networkx>=2.6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Image processing
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# Configuration management
PyYAML>=6.0
omegaconf>=2.1.0

# Progress bars and utilities
tqdm>=4.62.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Logging and monitoring
tensorboard>=2.7.0
wandb>=0.12.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.9.0
flake8>=4.0.0
isort>=5.9.0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0

# Type checking
mypy>=0.910

# Performance profiling
line_profiler>=3.3.0
memory_profiler>=0.60.0

# File I/O and data formats
h5py>=3.4.0
pickle5>=0.0.11

# Reproducibility
python-dotenv>=0.19.0

# Academic and research tools
bibtexparser>=1.2.0

# Optional GPU acceleration (uncomment if using CUDA)
# torch-geometric>=2.0.0
# torch-scatter>=2.0.0
# torch-sparse>=0.6.0
# torch-cluster>=1.5.0

# For semantic map processing
rasterio>=1.2.0
geopandas>=0.10.0

# For trajectory analysis
shapely>=1.8.0

# Development dependencies (optional)
pre-commit>=2.15.0
bandit>=1.7.0
